--- app/components/layout/BlurableContainer.js
@@ -1,9 +1,0 @@
-import { modalVisible } from "connectors";
-
-const BlurableContainer = ({ modalVisible, className, children }) => (
-  <div className={[className, modalVisible ? "blur" : ""].join(" ")}>
-    {children}
-  </div>
-);
-
-export default modalVisible(BlurableContainer);
