--- app/components/layout/DescriptionHeader.js
@@ -1,8 +1,0 @@
-const DescriptionHeader = ({ description, actionButton }) => (
-  <div className="description-header">
-    <div className="description-header-action-button">{actionButton}</div>
-    {description}
-  </div>
-);
-
-export default DescriptionHeader;
