+++ app/components/layout/DescriptionHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from "./DescriptionHeader";
