--- app/components/layout/StandaloneHeader.js
@@ -1,22 +1,0 @@
-import { TitleHeader } from "./TitleHeader";
-import DescriptionHeader from "./DescriptionHeader";
-
-const StandaloneHeader = ({
-  title,
-  description,
-  iconType,
-  actionButton
-}) => {
-  return (
-    <div className="standalone-page-header">
-      <TitleHeader
-        title={title}
-        iconType={iconType}
-        optionalButton={actionButton}
-      />
-      <DescriptionHeader description={description} />
-    </div>
-  );
-};
-
-export default StandaloneHeader;
