+++ app/components/layout/StandaloneHeader/StandaloneHeader.module.css
@@ -1,0 +1,32 @@
+.header {
+  background-color: var(--background-back-color);
+  padding: 43px 60px 0px 63px;
+  position: absolute;
+  top: 0px;
+  right: 0px;
+  left: 0px;
+  height: 114px;
+}
+
+@media screen and (max-width: 1179px) {
+  .header {
+    padding-left: 20px;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  /* XXX wassuuup here */
+  .tabbed-page-header > .standalone-page-header {
+    padding: 0;
+  }
+
+  .header {
+    height: initial;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    position: relative;
+    padding-bottom: 48px;
+    padding-top: 30px;
+  }
+}
