+++ app/style/Layout.less
@@ -95,8 +95,7 @@
   left: 0px;
 }
 
-.tabbed-page-header,
-.standalone-page-header {
+.tabbed-page-header {
   background-color: var(--background-back-color);
   padding: 43px 60px 0px 63px;
   position: absolute;
@@ -128,24 +127,7 @@
   background-color: var(--background-container);
 }
 
-.tabbed-page-header > .description-header,
-.standalone-page-header > .description-header {
-  white-space: pre-wrap;
-  color: var(--header-desc-lighter-color);
-  font-size: 13px;
-  float: left;
-  width: 100%;
-  margin-left: 40px;
-}
 
-.description-header > .description-header-action-button {
-  position: absolute;
-  right: 80px;
-}
-
-.description-header > .description-header-action-button > .button {
-  min-width: 76px;
-}
 
 .tabbed-page-body.governance {
   top: 157px;
@@ -354,7 +336,6 @@
     padding-left: 35px;
   }
   .tabbed-page-header,
-  .standalone-page-header,
   .standalone-page-body {
     padding-left: 20px;
   }
